Minor Refactoring Base test
