+++ src/test/java/org/example/puppy/base/BaseWebTest.java
@@ -4,7 +4,7 @@ import org.example.core.DriverHolder;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 
-public class BaseTest {
+public class BaseWebTest {
 
     @BeforeEach
     public void setup() {
