+++ src/test/java/org/example/puppy/login/EmptyCredentialsTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.login;
 
 import org.example.pages.LoginPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class EmptyCredentialsTest extends BaseTest {
+public class EmptyCredentialsTest extends BaseWebTest {
     private final LoginPage loginPage = new LoginPage();
 
     @Test
