+++ src/test/java/org/example/puppy/login/EmptyLoginTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.login;
 
 import org.example.pages.LoginPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class EmptyLoginTest extends BaseTest {
+public class EmptyLoginTest extends BaseWebTest {
     private static final String PASSWORD = "password";
     private final LoginPage loginPage = new LoginPage();
