+++ src/test/java/org/example/puppy/login/EmptyPasswordTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.login;
 
 import org.example.pages.LoginPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class EmptyPasswordTest extends BaseTest {
+public class EmptyPasswordTest extends BaseWebTest {
     private static final String USERNAME = "admin";
     private final LoginPage loginPage = new LoginPage();
