+++ src/test/java/org/example/puppy/login/LoginLayoutTest.java
@@ -2,10 +2,10 @@ package org.example.puppy.login;
 
 import org.example.pages.LoginPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
 
-public class LoginLayoutTest extends BaseTest {
+public class LoginLayoutTest extends BaseWebTest {
     private final LoginPage loginPage = new LoginPage();
