+++ src/test/java/org/example/puppy/login/LoginTest.java
@@ -3,10 +3,10 @@ package org.example.puppy.login;
 import org.example.pages.DashboardPage;
 import org.example.pages.LoginPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
 
-public class LoginTest extends BaseTest {
+public class LoginTest extends BaseWebTest {
     private static final String USERNAME = "admin";
     private static final String PASSWORD = "password";
     private final LoginPage loginPage = new LoginPage();
