+++ src/test/java/org/example/puppy/login/WrongCredentialsTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.login;
 
 import org.example.pages.LoginPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class WrongCredentialsTest extends BaseTest {
+public class WrongCredentialsTest extends BaseWebTest {
     private static final int RANDOM_CREDENTIALS_LENGHT = 6;
     private final LoginPage loginPage = new LoginPage();
