+++ src/test/java/org/example/puppy/login/WrongLoginTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.login;
 
 import org.example.pages.LoginPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class WrongLoginTest extends BaseTest {
+public class WrongLoginTest extends BaseWebTest {
     private static final int RANDOM_USERNAME_LENGHT = 6;
     private static final String PASSWORD = "password";
     private final LoginPage loginPage = new LoginPage();
