+++ src/test/java/org/example/puppy/registration/RegistrationEmptyPasswordFieldTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.registration;
 
 import org.example.pages.RegistrationPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class RegistrationEmptyPasswordFieldTest extends BaseTest {
+public class RegistrationEmptyPasswordFieldTest extends BaseWebTest {
 
     private static final String USERNAME = "Newaccount7";
     private final RegistrationPage registrationPage = new RegistrationPage();
