+++ src/test/java/org/example/puppy/registration/RegistrationEmptyRepeatPasswordFieldTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.registration;
 
 import org.example.pages.RegistrationPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class RegistrationEmptyRepeatPasswordFieldTest extends BaseTest {
+public class RegistrationEmptyRepeatPasswordFieldTest extends BaseWebTest {
 
     private static final String USERNAME = "Newaccount8";
     private static final String PASSWORD = "AbobaPuppy";
