+++ src/test/java/org/example/puppy/registration/RegistrationEmptyUsernameFieldTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.registration;
 
 import org.example.pages.RegistrationPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class RegistrationEmptyUsernameFieldTest extends BaseTest {
+public class RegistrationEmptyUsernameFieldTest extends BaseWebTest {
 
     private static final String PASSWORD = "AbobaPuppy";
     private final RegistrationPage registrationPage = new RegistrationPage();
