+++ src/test/java/org/example/puppy/registration/RegistrationShortPasswordTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.registration;
 
 import org.example.pages.RegistrationPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class RegistrationShortPasswordTest extends BaseTest {
+public class RegistrationShortPasswordTest extends BaseWebTest {
 
     private static final String USERNAME = "Newacconut4";
     private static final String PASSWORD = "Aboba";
