+++ src/test/java/org/example/puppy/registration/RegistrationUsernameALotOfCharactersTest.java
@@ -2,9 +2,9 @@ package org.example.puppy.registration;
 
 import org.example.pages.RegistrationPage;
 import org.junit.jupiter.api.Test;
-import org.example.puppy.base.BaseTest;
+import org.example.puppy.base.BaseWebTest;
 
-public class RegistrationUsernameALotOfCharactersTest extends BaseTest {
+public class RegistrationUsernameALotOfCharactersTest extends BaseWebTest {
     private static final int RANDOM_USERNAME_LENGHT = 151;
     private static final String PASSWORD = "AbobaPuppy";
     private final RegistrationPage registrationPage = new RegistrationPage();
